requirements.txt
APScheduler==3.10.4
awsebcli==3.20.10
blinker==1.6.2
botocore==1.31.85
Bottleneck==1.3.7
cement==2.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
contourpy==1.2.0
cryptography==42.0.8
cycler==0.11.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Mail==0.10.0
fonttools==4.51.0
idna==3.7
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jwt==1.3.1
kaleido==0.1.0
kiwisolver==1.4.4
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.4
numexpr==2.8.7
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
pathspec==0.10.1
patsy==0.5.6
Pillow==10.3.0
plotly==5.22.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
schedule==1.2.2
scikit-learn==1.4.2
scipy==1.13.1
semantic-version==2.8.5
six==1.16.0
smmap==5.0.1
statsmodels==0.14.2
tenacity==8.4.2
termcolor==1.1.0
threadpoolctl==3.5.0
tomli==2.0.1
tornado==6.4.1
tzlocal==5.2
unicodedata2==15.1.0
urllib3==1.26.19
wcwidth==0.1.9
Werkzeug==3.0.3
zipp==3.17.0
itsdangerous==2.2.0
boto3==1.31.85